+++ Panther/src/sema_to_pir/SemaToPIRData.hpp
@@ -83,7 +83,6 @@ namespace pcit::panther{
 				bool isNoReturn;
 				evo::SmallVector<Param> params;
 				evo::SmallVector<ReturnParam> return_params; // only used if they are out params
-				std::optional<pir::Expr> error_return_param;
 				std::optional<pir::Type> error_return_type;
 			};